backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
beautifulsoup4
requests
google-generativeai
